Refactor StitchHandler into specialized services
Decomposed `StitchHandler` (a "God Class") into three domain-specific services:
- `StitchIamService`: Handles IAM policy bindings.
- `StitchApiService`: Handles API enablement.
- `StitchConnectionService`: Handles connectivity checks.

`StitchHandler` now acts as a facade, orchestrating these services.
Also updated `PREVENT_CONFLICTS.md` to reflect the completion of this task and the previous `GcloudHandler` decomposition.

Tests passed locally.
